+++ toolchest/rpm/rpmvercmp.py
