+++ toolchest/strutil.py
