+++ tox.ini
@@ -19,4 +19,4 @@ passenv=HOME
 sitepackages = False
 deps = flake8
 commands =
-    flake8 --ignore=E501 setup.py rd_utils tests
+    flake8 --ignore=E501 setup.py toolchest tests
